api/pyproject.toml
[project]
name = "avaliar"
version = "0.0.0"
description = "Sistema de Gestão de Provas"
authors = [{ name = "Braz Souza", email = "mail@braz.one" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "fastapi[standard]>=0.116.1,<0.117.0",
    "poethepoet>=0.37.0,<0.38",
    "pydantic-settings>=2.0.0",
    "authx>=1.0.0",
    "sqlmodel>=0.0.27",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "alembic>=1.12.0,<2.0.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.poe.tasks]
start = "python -c \"import subprocess; import threading; import os; p1 = subprocess.Popen(['uvicorn', 'main:app', '--reload', '--host', '0.0.0.0', '--port', '8000']); p2 = subprocess.Popen(['npm', 'start'], cwd='./front'); p1.wait(); p2.wait()\""
api = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
front = {shell = "cd front && npm start"}
dev = {shell = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"}

[tool.poe.tasks.install-deps]
shell = "npm install --prefix ./front"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
